% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sggcode.R
\name{get_sgg_code_name_df}
\alias{get_sgg_code_name_df}
\title{A sgg_code function}
\usage{
get_sgg_code_name_df()
}
\description{
This function allows you to get sgg code data frame.
}
\examples{
get_sgg_code_name_df()
}
\keyword{sgg_code}
