% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sggcode.R
\name{get_sggcode}
\alias{get_sggcode}
\title{Get Sigungu Code Character String}
\usage{
get_sggcode(sgg_code_name_df, sidonamestr, sggnamestr)
}
\arguments{
\item{sgg_code_name_df}{sgg_code_name dataframe.}

\item{sidonamestr}{sido_name string (e.g. '서울특별시', '경기도').}

\item{sggnamestr}{sgg_name string (e.g. '강남구', '수원시').}
}
\value{
A character string of sgg_code
}
\description{
This function allows you to get sgg_code by passing ssg_name.
}
\examples{
get_sggcode(sgg_code_name_df, '서울특별시', '강남구')
}
\keyword{sgg_code}
\keyword{sgg_code,}
\keyword{sgg_name,}
\keyword{sido_name,}
