% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sggcode.R
\name{get_sidocode}
\alias{get_sidocode}
\title{A sgg_code function}
\usage{
get_sidocode(sgg_code_name_df, sidonamestr)
}
\arguments{
\item{sgg_code_name_df}{sgg_code_name dataframe.}

\item{sidonamestr}{sido_name string (e.g. '서울특별시', '경기도').}
}
\description{
This function allows you to get sido_code by passing sido_name.
}
\examples{
get_sidocode(sgg_code_name_df, '서울특별시')
}
\keyword{sgg_code,}
\keyword{sido_code}
\keyword{sido_name,}
